# Fix performance regression in `TimeWithZone#to_time`