--- a/activesupport/lib/active_support/time_with_zone.rb
+++ b/activesupport/lib/active_support/time_with_zone.rb
@@ -80,15 +80,9 @@ module ActiveSupport
 
     # Returns a <tt>Time</tt> instance of the simultaneous time in the system timezone.
     def localtime(utc_offset = nil)
-      utc.getlocal(utc_offset)
+      @localtime ||= utc.getlocal(utc_offset)
     end
     alias_method :getlocal, :localtime
-
-    def to_time_with_caching
-      @to_time ||= to_time_without_caching
-    end
-    alias_method :to_time_without_caching, :to_time
-    alias_method :to_time, :to_time_with_caching
 
     # Returns true if the current time is within Daylight Savings Time for the
     # specified time zone.