@clemensp 2016-07-18
We would like to have this back-ported to Rails 4.2 as well.

Ideally, it would be back-ported to Rails 4.1, where we are currently experiencing a big performance degradation, but I assume that it is not an option anymore.