@guilleiguaran 2016-07-19
Sorry, this can't be backported to Rails 4.1 but we will consider to do it to 4.2