@clemensp 2016-08-04
@rafaelfranca What would you say is an acceptable waiting period before asking for updates?  😃
I'm not sure what the standard wait time for Rails PRs is.

In any case, I'd like to see if the fix looks reasonable, so that I can replace some workarounds.