@pixeltrix 2016-08-05
@clemensp sorry, day job got in the way

@ryandv why the separate method and aliasing? Why not caching in the `to_time` method directly? e.g:

``` ruby
def to_time
  @to_time ||= utc.to_time
end
```

Also you may want to double check there's no way of mutating an AS::TWZ instance - I don't think so but best to be sure.