@ryandv 2016-08-05
@pixeltrix We were following existing conventions within ActiveSupport for overriding methods, as exemplified at https://github.com/rails/rails/blob/5-0-stable/activesupport/lib/active_support/core_ext/date/calculations.rb#L141

However, we are not attached to either approach and will happily follow your suggestion.

We did identify a mutable configuration option `to_time_preserves_timezone`, which could cause the `utc_offset` used to localize the value of `to_time` to fall out of sync with the memoized attribute. However, the intended usage of this option appears to be as an initialization configuration.

See https://github.com/rails/rails/commit/c9c5788a527b70d7f983e2b4b47e3afd863d9f48 for relevant documentation.