@pixeltrix 2016-08-05
@ryandv yes, I think `to_time_preserves_timezone` shouldn't be a problem. Thinking about it the caching probably needs to implemented in the `localtime` method since that's what `to_time` eventually calls, e.g:

``` ruby
def localtime(utc_offset = nil)
  @localtime ||= utc.getlocal(utc_offset)
end
alias_method :getlocal, :localtime
```

A quick check here shows all the Active Support tests pass