@ryandv 2016-08-23
As per your suggestions, we've removed method aliasing and moved the memoization
to `TimeWithZone#localtime`. Both our original changes and your suggested
revisions address the performance regression. However, our original
fix is directly backportable to `4-2-stable`. This is because `localtime`
is not along the codepath used by `Time#compare_with_coercion` in `4-2-stable`.

Compare the following:

* `4-2-stable`: https://github.com/rails/rails/blob/4-2-stable/activesupport/lib/active_support/time_with_zone.rb#L320
* `5-0-stable`: https://github.com/rails/rails/blob/5-0-stable/activesupport/lib/active_support/core_ext/date_and_time/compatibility.rb#L15
  * In particular, `5-0-stable` introduces `DateAndTime::Compatibility` which has `#to_time` delegate to `localtime`.

The following benchmark validates that the revised changeset still performs
as expected:

```rb
require 'active_support/time'
require 'benchmark/ips'

patchedTimeWithZone = Class.new(ActiveSupport::TimeWithZone) do
  def getlocal
    @getlocal ||= super
  end
end

utc = Time.utc(2000, 1, 1, 0)
time_zone = ActiveSupport::TimeZone['Eastern Time (US & Canada)']

twz = ActiveSupport::TimeWithZone.new(utc, time_zone)
twz2 = ActiveSupport::TimeWithZone.new(Time.utc(1999, 12, 31, 23, 59, 59), ActiveSupport::TimeZone['UTC'])

patched_twz = patchedTimeWithZone.new(utc, time_zone)
patched_twz2 = patchedTimeWithZone.new(Time.utc(1999, 12, 31, 23, 59, 59), ActiveSupport::TimeZone['UTC'])

Benchmark.ips do |x|
  x.report("TimeWithZone comparison - existing implementation") { twz <=> twz2 }
  x.report("TimeWithZone comparison - caching implementation") { patched_twz <=> patched_twz2 }
  x.compare!
end
```

The results, in rails-dev-box, compared against `58e0b1d`:

```
vagrant@rails-dev-box:/vagrant/rails$ bundle exec ruby cached_localtime_performance.rb
Warming up --------------------------------------
TimeWithZone comparison - existing implementation
                        25.925k i/100ms
TimeWithZone comparison - caching implementation
                        53.263k i/100ms
Calculating -------------------------------------
TimeWithZone comparison - existing implementation
                        446.104k (±12.5%) i/s -      2.204M in   5.043571s
TimeWithZone comparison - caching implementation
                          2.128M (± 5.7%) i/s -     10.599M in   4.998911s

Comparison:
TimeWithZone comparison - caching implementation:  2127709.3 i/s
TimeWithZone comparison - existing implementation:   446103.9 i/s - 4.77x slower
```